+++ src/neuroconv/tools/testing/__init__.py
@@ -8,6 +8,7 @@ from .mock_files import generate_path_expander_demo_ibl
 from .mock_interfaces import (
     MockBehaviorEventInterface,
     MockEventsInterface,
+    MockFiberPhotometryInterface,
     MockSpikeGLXNIDQInterface,
     MockRecordingInterface,
     MockImagingInterface,